Comments and filter posts
Added some functuanality to add comments to posts, but they are not yet displayed. I will add this later on if I have the time. They are stored on the id of the post, and I can access them through the api at this point.
Also added some html for filtering posts, but will add funct. for this next time.