profile
Added a more dynamic profile thats gets its data based on the user that is logged in. The profile image is getting a dropdown menu on click, so the user can watch / change / delete the image. This gets stored on the object on the server. This is not ready at this time.