Fixed login and logout issues
Fixed problem when user session didnt get cleared on logout. This is now ok.
When users log in, some tweaks have been made in terms of displaying correct posts on profile page.
Having some issues regarding displaying avatar in header when this is not set. is supposed to be a default image, but wont load, so it looks broken. setting a profile avatar on profile fixes this.