Posts storage fix
When user logs in to the website, the posts in the feed gets stored in sessionStorage. This only gets cleared on logout (or when closing tab).
Helps handle 429 error, but does not update the posts displayed while the user is logged in (say you create a new post, you will not see it in the "global" feed. Only on the profile. A relog fixes this